Entity, State 구조 일부 수정
Entity, State 구조 일부 수정 / Enemy 임시 추가

+++ Assets/01.Scripts/Combat/IDamageable.cs
@@ -11,5 +11,5 @@ public enum DamageType
 
 public interface IDamageable
 {
-    public void OnDamage(DamageType type, float damage);
+    public void OnDamage(DamageType type, int damage);
 }
